refact: extract mail and channel resolution logic into traits
Move toMail/resolveRecipientEmail into NotificationMailTrait (consistent
with existing NotificationSmsTrait, NotificationOneSignalTrait pattern).

Move channels/setChannels/getNotificationChannels/filterEnabledChannels
into NotificationChannelResolutionTrait to keep the main Notification
class focused on orchestration.
